[wip] Split `wallet-core-rs` crate into multiple ones
