--- rust/src/lib.rs
@@ -1,13 +1,0 @@
-// Copyright © 2017-2022 Trust Wallet.
-//
-// This file is part of Trust. The full Trust copyright notice, including
-// terms governing use, modification, and redistribution, is contained in the
-// file LICENSE at the root of the source code distribution tree.
-
-extern crate core;
-
-pub mod memory;
-pub mod hash;
-pub mod encoding;
-pub mod move_parser;
-pub mod starknet;
