+++ rust/tw_encoding/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "tw_encoding"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+base64 = "0.21.0"
+bs58 = "0.4.0"
+hex = "0.4.3"
+tw_memory = { path = "../tw_memory" }
