+++ rust/tw_hash/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "tw_hash"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+blake-hash = "0.4.1"
+blake2 = "0.10.6"
+blake2b-ref = "0.3.1"
+digest = "0.10.6"
+groestl = "0.10.1"
+hex = "0.4.3"
+hmac = "0.12.1"
+ripemd = "0.1.3"
+sha1 = "0.10.5"
+sha2 = "0.10.6"
+sha3 = "0.10.6"
+tw_memory = { path = "../tw_memory" }
