+++ rust/tw_hash/src/hash_wrapper.rs
