+++ rust/tw_hash/src/lib.rs
