+++ rust/tw_memory/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "tw_memory"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
