+++ rust/tw_starknet/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "tw_starknet"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+hex = "0.4.3"
+starknet-crypto = "0.1.0"
+starknet-ff = "0.1.0"
+starknet-signers = "0.1.0"
+tw_memory = { path = "../tw_memory" }
