+++ rust/wallet_core_rs/src/lib.rs
@@ -1,0 +1,1 @@
+
